+++ UnitTests/BraintreeApplePayTests/BTApplePay_Tests.swift
@@ -22,7 +22,7 @@ class BTApplePay_Tests: XCTestCase {
         let applePayClient = BTApplePayClient(apiClient: mockClient)
 
         let expectation = self.expectation(description: "Callback invoked")
-        applePayClient.paymentRequest { (paymentRequest, error) in
+        applePayClient.makePaymentRequest { (paymentRequest, error) in
             guard let error = error as NSError? else {return}
             XCTAssertEqual(error.domain, BTApplePayError.errorDomain)
             XCTAssertEqual(error.code, BTApplePayError.unsupported.rawValue)
@@ -38,7 +38,7 @@ class BTApplePay_Tests: XCTestCase {
         let applePayClient = BTApplePayClient(apiClient: mockClient)
 
         let expectation = self.expectation(description: "Callback invoked")
-        applePayClient.paymentRequest { (paymentRequest, error) in
+        applePayClient.makePaymentRequest { (paymentRequest, error) in
             guard let error = error as NSError? else {return}
             XCTAssertEqual(error.domain, BTApplePayError.errorDomain)
             XCTAssertEqual(error.code, BTApplePayError.unsupported.rawValue)
@@ -62,7 +62,7 @@ class BTApplePay_Tests: XCTestCase {
         let applePayClient = BTApplePayClient(apiClient: mockClient)
 
         let expectation = self.expectation(description: "Callback invoked")
-        applePayClient.paymentRequest { (paymentRequest, error) in
+        applePayClient.makePaymentRequest { (paymentRequest, error) in
             guard let paymentRequest = paymentRequest else {
                 XCTFail()
                 return
@@ -88,7 +88,7 @@ class BTApplePay_Tests: XCTestCase {
         let applePayClient = BTApplePayClient(apiClient: mockClient)
 
         let expectation = self.expectation(description: "Callback invoked")
-        applePayClient.paymentRequest { (paymentRequest, error) in
+        applePayClient.makePaymentRequest { (paymentRequest, error) in
             guard let paymentRequest = paymentRequest else {
                 XCTFail()
                 return
@@ -109,7 +109,7 @@ class BTApplePay_Tests: XCTestCase {
         let applePayClient = BTApplePayClient(apiClient: mockClient)
         mockClient.cannedConfigurationResponseBody = BTJSON(value: ["applePay" : ["status" : "production"]])
 
-        let paymentRequest = try await applePayClient.paymentRequest()
+        let paymentRequest = try await applePayClient.makePaymentRequest()
         XCTAssertNotNil(paymentRequest)
     }
 
@@ -118,7 +118,7 @@ class BTApplePay_Tests: XCTestCase {
         mockClient.cannedConfigurationResponseBody = BTJSON(value: ["applePay" : ["status" : "off"]])
 
         do {
-            let _ = try await applePayClient.paymentRequest()
+            let _ = try await applePayClient.makePaymentRequest()
         } catch {
             let error = error as NSError
             XCTAssertNotNil(error)